Partial Reorganization and Simplification
Previously, javascript partials accepted and returned a KVP. They now
accept the parameters to the partial (object, array, or simple value)
and return a value.

Previously, we normalized the parameters to the partial to be an
object, and we added a key and optionally a value parameter. We no
longer do that. The parameters to a partial may be an object, or it
may be an array or simple value.