Remove variant-with-alternate-index
Chose to express the realm index is an HTTP header when there are more
than one variants present. The thinking is that the user agent can
take on knowledge of this case and provide a link or some other
visualization to indicate that other variants exist that a user could
select.