Deal with values that look like lynx but aren't
Added code to handle situations where a value has a spec but no
hints. The value is now not considered a lynx value and a warning
is generated about the value that appears to be lynx but isn't.

Changed incompatible sections to be a debug level message instead of
a warn level message.