Better error content when generating lynx
Instead of throwing errors out of the lynx generation process we capture them and create a lynx document with the error information in the same realm of the document that was requested. This is to surface more meaningful error information to the user as well as keep scoped content from taking over the entire view when an error occurs.
- Generates specific content for YAML parse failures
- Generates specific content for JSON lint failures during data binding.
- Appended data file name to error message when resolving data files so that name was available in the error document.