Issue 86 (#89)
* Resolve flattening issue generating invalid JSON

This issue was resolved by handling implicit value components in
the lynx partial. There is probably a better phrase than implicit value
component, but it's the best I have right now.
An implicit value component is created for a partial with a wildcard
placeholder when the parameter passed to the partial is not an object.
Consider as an example the `text` partial.
```yaml
>lynx:
  spec.hints: ["text"]
  *~:
```
When this partial is called with a string parameter like "Hello World"
(e.g. key>text: "Hello World") we need to put "Hello World" somewhere in
the new parameter set that is being passed to the lynx partial. We are
currently placing that parameter on an object with a single empty key.
```javascript
{ "": "Hello World" }
```
So the resulting parameters passed to the lynx partial are
```javascript
{ spec.hints: ["text"], "": "Hello world" }
```
We do this because the following is invalid
```javascript
{ spec.hints: ["text"], "Hello world" }
```
By adding knowledge of this concept to the lynx partial it can do the
right thing and treat the empty key similar to if the value component
was explicitly stated.

* Resolve flattening issue generating invalid JSON

This issue was resolved by handling implicit value components in
the lynx partial. There is probably a better phrase than implicit value
component, but it's the best I have right now.
An implicit value component is created for a partial with a wildcard
placeholder when the parameter passed to the partial is not an object.
Consider as an example the `text` partial.
```yaml
>lynx:
  spec.hints: ["text"]
  *~:
```
When this partial is called with a string parameter like "Hello World"
(e.g. key>text: "Hello World") we need to put "Hello World" somewhere in
the new parameter set that is being passed to the lynx partial. We are
currently placing that parameter on an object with a single empty key.
```javascript
{ "": "Hello World" }
```
So the resulting parameters passed to the lynx partial are
```javascript
{ spec.hints: ["text"], "": "Hello world" }
```
We do this because the following is invalid
```javascript
{ spec.hints: ["text"], "Hello world" }
```
By adding knowledge of this concept to the lynx partial it can do the
right thing and treat the empty key similar to if the value component
was explicitly stated.

* Fix content issue with test

* Move knowledge of non object parameters to module

- Ability to specify where to place non object partial parameters. Used
this ability to configure that those values are placed in the `value`
compononent for lynx-docs.
- Change to make sure that we can activate and deactivate flattening
from the CLI and .lynxdocsrc run control file.