Resolves 78. Resolves 91. Resolves 93. (#94)
* Resolve 78, Resolve 91, refactor server

This change was specifically to resolve issue 78. During the course of the work it became apparent is was a good opportunity to provide some additional organization to the server and resolve issue #91 at the same time.

Also folded in the change to do better reporting of jsmodule resolution failures since I was refactoring that module.