Partial processing for in-memory data
This change allows in memory data representations for variants to participate in partial processing before being bound to the template. This support scenarios where JavaScript variants load data files with partial tokens in them, perform some small amount of mutation, and then want the rest of the token syntax processing to happen as it would if it was loaded from a file.