Refactor partial processing to resolve 108 (#110)
Resolves #101
Resolves #108

* Provide options context to partials and data

This change adds the running options context when calling partial or data modules. When running from the `start` command, the `req` and `res` objects for the current web request are appended to options. Modules can inspect the options context and implement custom logic.

* Refactor partial processing to resolve 108

* Pass options to partial expansion

* Allow multiple usage of partial parameter

* Partial processing for in-memory data

This change allows in memory data representations for variants to participate in partial processing before being bound to the template. This supports scenarios where JavaScript variants load data files with partial tokens in them, perform some small amount of mutation, and then want the rest of the syntax processing to happen as it would if it was loaded from a file.